+++ scalar_multiplication.c
@@ -1,4 +1,4 @@
-int **scalar_multiplication(int rows, int cols, double scalar, double matrix[rows][cols]) {
+double **scalar_multiplication(int rows, int cols, double scalar, double matrix[rows][cols]) {
     double **result = malloc(rows * sizeof(double *));
     for (int i = 0; i < rows; i++) {
         result[i] = malloc(cols * sizeof(double));
